backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gerenciador-estoque-carros"
version = "0.1.0"
description = "Sistema de Gerenciamento de Estoque"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "psycopg2-binary"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[tool.setuptools]
packages = ["src"]
